Update `getQuestWithSteps` parameter type and add test data generator:
- Introduce `TestDataGenerator` for creating mock entities (`PlayerEntity`, `MapQuestEntity`, `QuestStepEntity`) for testing purposes.

+++ app/src/main/java/ee/ut/cs/tartu_explorer/core/data/local/dao/MapQuestDao.kt
@@ -20,7 +20,7 @@ interface MapQuestDao {
 
     @Transaction
     @Query("SELECT * FROM map_quest WHERE id = :questId")
-    suspend fun getQuestWithSteps(questId: Long): MapQuestWithSteps?
+    suspend fun getQuestWithSteps(questId: Int): MapQuestWithSteps?
 
     @Transaction
     @Query("SELECT * FROM map_quest")
